+++ R/query_device.R
@@ -23,7 +23,7 @@ query_ap_device <- function(device_ip, query) {
   req <- request(url)
   resp <- req |> req_perform()
   if (resp_is_error(resp)) {
-    cli::cli_abort(c("Connection to device {.var {device_ip}} raise an error : ",
+    cli::cli_abort(c("Connection to {.var {device_ip}} device{?s} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
 
   } else if (inherits(resp, "httr2_failure")) {
@@ -63,12 +63,12 @@ query_ap_devices <- function(device_ip, query) {
   response_is_error <- map_lgl(resp, inherits, "httr2_failure")
 
   if (all(response_is_error)) {
-    cli::cli_abort("Connection to all devices raised an error.")
+    cli::cli_abort("Connection to all of {.val {unique(device_ip)}} device{?s} raised an error.")
   }
 
   if (any(response_is_error)) {
     cli::cli_warn(c(
-      "Connection to device {.var {device_ip[response_is_error]}} raise an error : ",
+      "Connection to {.var {device_ip[response_is_error]}} device{?s} raise an error : ",
       "{map(which(response_is_error), ~resp_status(resp[[.x]]))} {map(which(response_is_error), ~resp_status_desc(resp[[.x]]))}."
       ))
   }
@@ -108,7 +108,7 @@ query_enphaseenvoy_device <- function(device_ip = "enphase.local", query, userna
   req <- request(url) |> req_auth_basic(username, password)
   resp <- req |> req_perform()
   if (resp_is_error(resp)) {
-    cli::cli_abort(c("Connection to device {.var {device_ip}} raise an error : ",
+    cli::cli_abort(c("Connection to {.var {device_ip}} device{?s} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
 
   } else if (inherits(resp, "httr2_failure")) {
@@ -146,7 +146,7 @@ query_enphaseenergy_device <- function(device_ip = "enphase.local", query, usern
   req <- request(url) |> req_auth_basic(username, password)
   resp <- req |> req_perform()
   if (resp_is_error(resp)) {
-    cli::cli_abort(c("Connection to device {.var {device_ip}} raise an error : ",
+    cli::cli_abort(c("Connection to {.var {device_ip}} device{?s} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
 
   } else {
@@ -155,7 +155,7 @@ query_enphaseenergy_device <- function(device_ip = "enphase.local", query, usern
     if (length(info_lst[["data"]]) >= 3) {
       cbind(device_id = device_ip, as.data.frame(info_lst[["data"]]))
     } else {
-      cli::cli_abort(c("the Enphase device {.var {device_ip}} is not supported"))
+      cli::cli_abort(c("the Enphase {.var {device_ip}} device is not supported"))
     }
   }
 }
@@ -190,7 +190,7 @@ query_fronius_device <- function(device_ip = "fronius.local", query, username = 
   req <- request(url) |> req_auth_basic(username, password)
   resp <- req |> req_perform()
   if (resp_is_error(resp)) {
-    cli::cli_abort(c("Connection to device {.var {device_ip}} raise an error : ",
+    cli::cli_abort(c("Connection to {.var {device_ip}} device{?s} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
 
   } else if (inherits(resp, "httr2_failure")) {
@@ -202,7 +202,7 @@ query_fronius_device <- function(device_ip = "fronius.local", query, username = 
     if (info_lst[["Head"]][["Status"]][["Code"]] == 0) {
       cbind(device_id = device_ip, last_report = info_lst$Head$Timestamp, as.data.frame(info_lst$Body$Data))
     } else {
-      cli::cli_abort(c("the Fronius device {.var {device_ip}} does not have the correct Metering setup"))
+      cli::cli_abort("The {.var {device_ip[incorrect_status_code]}} Fronius device{?s} {?don't/doesn't} have the correct Metering setup")
     }
   }
 }
@@ -237,12 +237,12 @@ query_fronius_devices <- function(device_ip = c("fronius.local"), query, usernam
 
   response_is_error <- map_lgl(resp, inherits, "httr2_failure")
   if (all(response_is_error)) {
-    cli::cli_abort("Connection to all devices raised an error.")
+    cli::cli_abort("Connection to all of {.val {unique(device_ip)}} device{?s} raised an error.")
   }
 
   if (any(response_is_error)) {
     cli::cli_warn(c(
-      "Connection to device {.var {device_ip[response_is_error]}} raise an error : ",
+      "Connection to {.val {device_ip[response_is_error]}} device{?s} raise an error : ",
       "{map(which(response_is_error), ~resp_status(resp[[.x]]))} {map(which(response_is_error), ~resp_status_desc(resp[[.x]]))}."
     ))
   }
@@ -251,7 +251,7 @@ query_fronius_devices <- function(device_ip = c("fronius.local"), query, usernam
   incorrect_status_code <- map_lgl(info_lst, ~.x[["Head"]][["Status"]][["Code"]] != 0)
 
   if (any(incorrect_status_code)) {
-    cli::cli_warn("the Fronius device {.var {device_ip[incorrect_status_code]}} does not have the correct Metering setup")
+    cli::cli_warn("The {.var {device_ip[incorrect_status_code]}} Fronius device{?s} {?don't/doesn't} have the correct Metering setup")
   }
 
   map2_dfr(device_ip[!response_is_error][!incorrect_status_code], info_lst[!incorrect_status_code],